+++ Source/AtCoderAnalyzer/CreateOperators/Specified/ComparerEnumerateMember.cs
@@ -5,7 +5,7 @@ using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
 
-namespace AtCoderAnalyzer.CreateOperators
+namespace AtCoderAnalyzer.CreateOperators.Specified
 {
     internal class ComparerEnumerateMember : EnumerateMember
     {
