+++ Source/AtCoderAnalyzer/CreateOperators/Specified/DivisionEnumerateMember.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 
-namespace AtCoderAnalyzer.CreateOperators
+namespace AtCoderAnalyzer.CreateOperators.Specified
 {
     internal class DivisionEnumerateMember : OperatorEnumerateMember
     {
