+++ Source/AtCoderAnalyzer/CreateOperators/Specified/MultiplicationEnumerateMember.cs
@@ -4,7 +4,7 @@ using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
 
-namespace AtCoderAnalyzer.CreateOperators
+namespace AtCoderAnalyzer.CreateOperators.Specified
 {
     internal class MultiplicationEnumerateMember : OperatorEnumerateMember
     {
