+++ Source/AtCoderAnalyzer/CreateOperators/Specified/ShiftEnumerateMember.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 
-namespace AtCoderAnalyzer.CreateOperators
+namespace AtCoderAnalyzer.CreateOperators.Specified
 {
     internal class ShiftEnumerateMember : OperatorEnumerateMember
     {
