+++ Source/AtCoderAnalyzer/CreateOperators/Specified/SubtractEnumerateMember.cs
@@ -1,17 +1,17 @@
 ﻿using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 
-namespace AtCoderAnalyzer.CreateOperators
+namespace AtCoderAnalyzer.CreateOperators.Specified
 {
-    internal class AdditionEnumerateMember : OperatorEnumerateMember
+    internal class SubtractEnumerateMember : OperatorEnumerateMember
     {
-        internal AdditionEnumerateMember(ITypeSymbol typeSymbol) : base(typeSymbol) { }
+        internal SubtractEnumerateMember(ITypeSymbol typeSymbol) : base(typeSymbol) { }
 
         protected override SyntaxKind? GetSyntaxKind(IMethodSymbol symbol)
             => symbol switch
             {
                 { Parameters: { Length: not 2 } } => null,
-                { Name: "Add" } => SyntaxKind.AddExpression,
+                { Name: "Subtract" } => SyntaxKind.SubtractExpression,
                 _ => null,
             };
     }
