+++ Source/AtCoderAnalyzer/CreateOperators/Specified/UnaryNumEnumerateMember.cs
@@ -3,7 +3,7 @@ using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
 
-namespace AtCoderAnalyzer.CreateOperators
+namespace AtCoderAnalyzer.CreateOperators.Specified
 {
     internal class UnaryNumEnumerateMember : EnumerateMember
     {
